[RDE-431] Move Whisper and Storage calls into dedicated clients, remove sensors, move controllers to loop
