+++ src/controllerHost.ts
@@ -1,7 +1,7 @@
 import { Whisper } from './whisper';
 import { PluginHost } from './pluginHost';
 
-export interface ControllerHost extends PluginHost {
+export interface ControllerHost {
   /**
    * @returns - Promise resolving with the unique ID of the generated whisper.
    */
