+++ src/loopPlugin.ts
@@ -18,11 +18,4 @@ export interface LoopPlugin<THost> {
    * All plugin activity should stop when this is called.
    */
   stop(): void;
-
-  /**
-   * The host will send events to the plugin by calling this function.
-   *
-   * @param event
-   */
-  onEvent(event: PluginEvent): void;
 }
