+++ src/serve.ts
@@ -1,24 +1,12 @@
-import { Controller } from './controller';
+import { Loop } from './loop';
 import ControllerPlugin from './controllerPlugin';
-import { Sensor } from './sensor';
-import SensorPlugin from './sensorPlugin';
 
 /**
  * Takes a controller implementation and launches the plugin.
  *
  * @param controller - The controller implementation.
  */
-export function serveControllerPlugin(controller: Controller): void {
+export function serveControllerPlugin(controller: Loop): void {
   const plugin = new ControllerPlugin(controller);
-  plugin.serve();
-}
-
-/**
- * Takes a sensor implementation and launches the plugin.
- *
- * @param sensor - The sensor implementation.
- */
-export function serveSensorPlugin(sensor: Sensor): void {
-  const plugin = new SensorPlugin(sensor);
   plugin.serve();
 }
