+++ src/whisperGrpcHostClient.test.ts
@@ -1,0 +1,156 @@
+import { mocked } from 'ts-jest/utils';
+import Services from './proto/whisper_grpc_pb';
+import Messages from './proto/whisper_pb';
+import WhisperGrpcHostClient from './whisperGrpcHostClient';
+import { ConnInfo } from './proto/broker_pb';
+
+jest.mock('./proto/ldk_pb');
+jest.mock('./proto/ldk_grpc_pb');
+
+const WHISPER_ID = '1234-abcd';
+const hostClient = mocked(Services.WhisperClient);
+
+type CallbackHandlerFunc<TRequest = any, TResponse = any> = (
+  request: TRequest,
+  callback: (err: Error | null, response: TResponse) => void,
+) => void;
+
+describe('WhisperGrpcHostClient', () => {
+  let subject: WhisperGrpcHostClient;
+  let connInfo: ConnInfo.AsObject;
+  let waitForReadyMock: jest.Mock;
+  let emitWhisperMock: jest.Mock;
+  let updateWhisperMock: jest.Mock;
+
+  function createCallbackHandler(response?: any): CallbackHandlerFunc {
+    return (request, callback) => {
+      callback(null, response);
+    };
+  }
+  beforeEach(() => {
+    jest.resetAllMocks();
+    subject = new WhisperGrpcHostClient();
+    connInfo = {
+      address: 'a',
+      serviceId: 1,
+      network: 'n',
+    };
+    waitForReadyMock = jest.fn().mockImplementation(createCallbackHandler());
+    hostClient.mockImplementation(() => {
+      return {
+        emitWhisper: emitWhisperMock,
+        updateWhisper: updateWhisperMock,
+        waitForReady: waitForReadyMock,
+      } as any;
+    });
+  });
+  describe('#connect', () => {
+    it('instantiates a new host client and waits for it to be ready', async () => {
+      await expect(subject.connect(connInfo)).resolves.toBe(undefined);
+    });
+  });
+  describe('#emitWhisper', () => {
+    describe('when initialized', () => {
+      beforeEach(async () => {
+        emitWhisperMock = jest
+          .fn()
+          .mockImplementation(
+            createCallbackHandler({ getId: () => WHISPER_ID }),
+          );
+        await subject.connect(connInfo);
+        await subject.emitWhisper({
+          markdown: 'abc',
+          icon: 'ok',
+          label: 'Hey',
+        });
+      });
+      it('should call the client with a whisper message', async () => {
+        const whisperRequest = mocked(Messages.WhisperNewRequest).mock
+          .instances[0];
+        expect(emitWhisperMock).toHaveBeenCalledWith(
+          whisperRequest,
+          expect.anything(),
+        );
+      });
+      it('should set the style to default', () => {
+        const style = mocked(Messages.WhisperStyle).mock.instances[0];
+        expect(style.setBackgroundcolor).toHaveBeenCalledWith('#fff');
+        expect(style.setPrimarycolor).toHaveBeenCalledWith('#666');
+        expect(style.setHighlightcolor).toHaveBeenCalledWith('#651fff');
+      });
+      it('should create the whisper properly', () => {
+        const whisper = mocked(Messages.WhisperMsg).mock.instances[0];
+        expect(whisper.setMarkdown).toHaveBeenCalledWith('abc');
+        expect(whisper.setLabel).toHaveBeenCalledWith('Hey');
+        expect(whisper.setIcon).toHaveBeenCalledWith('ok');
+      });
+      it('should return the ID from the whisper', async () => {
+        await expect(
+          subject.emitWhisper({ markdown: 'a', label: 'a', icon: 'a' }),
+        ).resolves.toBe(WHISPER_ID);
+      });
+    });
+    describe('before connected', () => {
+      it('should throw an error', async () => {
+        await expect(
+          subject.emitWhisper({ markdown: 'a', label: 'a', icon: 'a' }),
+        ).rejects.toThrow('Accessing client before connected');
+      });
+    });
+  });
+  describe('#updatetWhisper', () => {
+    describe('when initialized', () => {
+      beforeEach(async () => {
+        updateWhisperMock = jest
+          .fn()
+          .mockImplementation(createCallbackHandler());
+        await subject.connect(connInfo);
+        await subject.updateWhisper(WHISPER_ID, {
+          markdown: 'abc',
+          icon: 'ok',
+          label: 'Hey',
+        });
+      });
+      it('should call the client with a whisper message', async () => {
+        const whisperRequest = mocked(Messages.WhisperUpdateRequest).mock
+          .instances[0];
+        expect(updateWhisperMock).toHaveBeenCalledWith(
+          whisperRequest,
+          expect.anything(),
+        );
+      });
+      it('should set the style to default', () => {
+        const style = mocked(Messages.WhisperStyle).mock.instances[0];
+        expect(style.setBackgroundcolor).toHaveBeenCalledWith('#fff');
+        expect(style.setPrimarycolor).toHaveBeenCalledWith('#666');
+        expect(style.setHighlightcolor).toHaveBeenCalledWith('#651fff');
+      });
+      it('should create the whisper properly', () => {
+        const whisper = mocked(Messages.WhisperMsg).mock.instances[0];
+        expect(whisper.setMarkdown).toHaveBeenCalledWith('abc');
+        expect(whisper.setLabel).toHaveBeenCalledWith('Hey');
+        expect(whisper.setIcon).toHaveBeenCalledWith('ok');
+      });
+      it('should resolve successfully', async () => {
+        await expect(
+          subject.updateWhisper(WHISPER_ID, {
+            markdown: 'a',
+            label: 'a',
+            icon: 'a',
+          }),
+        ).resolves;
+      });
+    });
+    describe('before connected', () => {
+      it('should throw an error', async () => {
+        await expect(
+          subject.updateWhisper(WHISPER_ID, {
+            markdown: 'a',
+            label: 'a',
+            icon: 'a',
+          }),
+        ).rejects.toThrow('Accessing client before connected');
+      });
+    });
+  });
+});
